silence more compiler warnings

+++ pluginapi.c
@@ -27,7 +27,7 @@ int NSISCALL popstringn(LPTSTR str, int maxlen)
   stack_t *th;
   if (!g_stacktop || !*g_stacktop) return 1;
   th=(*g_stacktop);
-  if (str) lstrcpyn(str,th->text,maxlen?maxlen:g_stringsize);
+  if (str) lstrcpyn(str,th->text,maxlen?maxlen:(int)g_stringsize);
   *g_stacktop = th->next;
   GlobalFree((HGLOBAL)th);
   return 0;
@@ -67,7 +67,7 @@ int NSISCALL PopStringA(LPSTR ansiStr)
 
 int NSISCALL PopStringNA(LPSTR ansiStr, int maxlen)
 {
-   int realLen = maxlen ? maxlen : g_stringsize;
+   int realLen = maxlen ? maxlen : (int)g_stringsize;
    LPWSTR wideStr = (LPWSTR) GlobalAlloc(GPTR, realLen*sizeof(WCHAR));
    int rval = popstringn(wideStr, realLen);
    WideCharToMultiByte(CP_ACP, 0, wideStr, -1, ansiStr, realLen, NULL, NULL);
@@ -117,7 +117,7 @@ int NSISCALL PopStringW(LPWSTR wideStr)
 
 int NSISCALL PopStringNW(LPWSTR wideStr, int maxlen)
 {
-   int realLen = maxlen ? maxlen : g_stringsize;
+   int realLen = maxlen ? maxlen : (int)g_stringsize;
    LPSTR ansiStr = (LPSTR) GlobalAlloc(GPTR, realLen);
    int rval = popstringn(ansiStr, realLen);
    MultiByteToWideChar(CP_ACP, 0, ansiStr, -1, wideStr, realLen);
